// generated/models/LoginUserVO.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type LoginUserVO = {
    account?: string;
    avatar?: string;
    createTime?: string;
    email?: string;
    id?: number;
    phone?: string;
    profile?: string;
    role?: string;
    updateTime?: string;
    username?: string;
};